requirements.txt
pyrogram
request
sqlite3
